ENH: Renamed 'Deformation' by 'Displacement'
To be consistent with the ITK naming, every occurence of
the word 'Deformation' has been replaced by 'Displacement'
