+++ include/itkFixedPointInverseDisplacementFieldImageFilter.h
@@ -15,8 +15,8 @@
  *  limitations under the License.
  *
  *=========================================================================*/
-#ifndef itkFixedPointInverseDeformationFieldImageFilter_h
-#define itkFixedPointInverseDeformationFieldImageFilter_h
+#ifndef itkFixedPointInverseDisplacementFieldImageFilter_h
+#define itkFixedPointInverseDisplacementFieldImageFilter_h
 
 
 #include "itkImageToImageFilter.h"
@@ -29,43 +29,43 @@
 namespace itk
 {
 
-/** \class FixedPointInverseDeformationFieldImageFilter
- * \brief Computes the inverse of a deformation field using a fixed point iteration scheme.
+/** \class FixedPointInverseDisplacementFieldImageFilter
+ * \brief Computes the inverse of a Displacement field using a fixed point iteration scheme.
  *
- * FixedPointInverseDeformationFieldImageFilter takes a deformation field as input and
- * computes the deformation field that is its inverse. If the input deformation
+ * FixedPointInverseDisplacementFieldImageFilter takes a Displacement field as input and
+ * computes the Displacement field that is its inverse. If the input Displacement
  * field was mapping coordinates from a space A into a space B, the output of
  * this filter will map coordinates from the space B into the space A.
  *
- * To compute the inverse of the given deformation field, the fixed point algorithm by
+ * To compute the inverse of the given Displacement field, the fixed point algorithm by
  * Mingli Chen, Weiguo Lu, Quan Chen, Knneth J. Ruchala and Gusavo H. Olivera
  * described in the paper
- * "A simple fixed-point approach to invert a deformation field",
+ * "A simple fixed-point approach to invert a Displacement field",
  * Medical Physics, vol. 35, issue 1, p. 81,
  * is applied.
  *
  * \author Marcel Lüthi, Computer Science Department, University of Basel
  *
  * \ingroup ImageToImageFilter
- * \ingroup FixedPointInverseDeformationField
+ * \ingroup FixedPointInverseDisplacementField
  */
 
 template < class TInputImage, class TOutputImage >
-class ITK_EXPORT FixedPointInverseDeformationFieldImageFilter :
+class ITK_EXPORT FixedPointInverseDisplacementFieldImageFilter :
     public ImageToImageFilter<TInputImage,TOutputImage>
 {
 public:
   /** Standard class typedefs. */
-  typedef FixedPointInverseDeformationFieldImageFilter   Self;
-  typedef ImageToImageFilter<TInputImage,TOutputImage>   Superclass;
-  typedef SmartPointer<Self>                             Pointer;
-  typedef SmartPointer<const Self>                       ConstPointer;
+  typedef FixedPointInverseDisplacementFieldImageFilter   Self;
+  typedef ImageToImageFilter<TInputImage,TOutputImage>    Superclass;
+  typedef SmartPointer<Self>                              Pointer;
+  typedef SmartPointer<const Self>                        ConstPointer;
 
   /** Method for creation through the object factory. */
   itkNewMacro(Self);
 
   /** Run-time type information (and related methods). */
-  itkTypeMacro(FixedPointInverseDeformationFieldImageFilter, ImageToImageFilter);
+  itkTypeMacro(FixedPointInverseDisplacementFieldImageFilter, ImageToImageFilter);
 
 
   /** Some typedefs. */
@@ -135,8 +135,8 @@ public:
 #endif
 
 protected:
-  FixedPointInverseDeformationFieldImageFilter();
-  ~FixedPointInverseDeformationFieldImageFilter() {}
+  FixedPointInverseDisplacementFieldImageFilter();
+  ~FixedPointInverseDisplacementFieldImageFilter() {}
 
   void PrintSelf(std::ostream& os, Indent indent) const;
 
@@ -146,7 +146,7 @@ protected:
 
 
 private:
-  FixedPointInverseDeformationFieldImageFilter(const Self&); //purposely not implemented
+  FixedPointInverseDisplacementFieldImageFilter(const Self&); //purposely not implemented
   void operator=(const Self&); //purposely not implemented
 
   OutputImageSizeType                      m_Size;              // Size of the output image
@@ -158,6 +158,6 @@ private:
 
 } // end namespace itk
 
-#include "itkFixedPointInverseDeformationFieldImageFilter.hxx"
+#include "itkFixedPointInverseDisplacementFieldImageFilter.hxx"
 
 #endif
