+++ itk-module.cmake
@@ -1,12 +1,12 @@
-set(DOCUMENTATION "FixedPointInverseDeformationField takes a deformation field as input and
-computes the deformation field that is its inverse. If the input deformation
+set(DOCUMENTATION "FixedPointInverseDisplacementField takes a Displacement field as input and
+computes the Displacement field that is its inverse. If the input Displacement
 field was mapping coordinates from a space A into a space B, the output of
 this filter will map coordinates from the space B into the space A.
 
-To compute the inverse of the given deformation field, the fixed point algorithm by
+To compute the inverse of the given Displacement field, the fixed point algorithm by
 Mingli Chen, Weiguo Lu, Quan Chen, Knneth J. Ruchala and Gusavo H. Olivera
 described in the paper
-\"A simple fixed-point approach to invert a deformation field\",
+\"A simple fixed-point approach to invert a Displacement field\",
 Medical Physics, vol. 35, issue 1, p. 81,
 is applied.
 
@@ -14,7 +14,7 @@ author Marcel Lüthi, Computer Science Department, University of Basel
 ")
 
 
-itk_module(FixedPointInverseDeformationField
+itk_module(FixedPointInverseDisplacementField
   DEPENDS
     ITKCommon
     ITKImageGrid
