+++ test/Docker/run.sh
@@ -4,6 +4,6 @@ script_dir="`cd $(dirname $0); pwd`"
 
 docker run \
   --rm \
-  -v $script_dir/../..:/usr/src/FixedPointInverseDeformationField \
-    insighttoolkit/fixedpointinversedeformationfield-test \
-      /usr/src/FixedPointInverseDeformationField/test/Docker/test.sh
+  -v $script_dir/../..:/usr/src/FixedPointInverseDisplacementField \
+    insighttoolkit/fixedpointinverseDisplacementfield-test \
+      /usr/src/FixedPointInverseDisplacementField/test/Docker/test.sh
